# asm/amd64: optimizes register3bits via array