--- a/internal/asm/amd64/impl_1_test.go
+++ b/internal/asm/amd64/impl_1_test.go
@@ -531,8 +531,8 @@ func TestAssemblerImpl_encodeNoneToNone(t *testing.T) {
 		{inst: CQO, exp: []byte{0x48, 0x99}},
 		{inst: NOP, exp: nil},
 		{inst: RET, exp: []byte{0xc3}},
-		{inst: REPMOVSQ, exp: []byte{0xf3, RexPrefixW, 0xa5}},
-		{inst: REPSTOSQ, exp: []byte{0xf3, RexPrefixW, 0xab}},
+		{inst: REPMOVSQ, exp: []byte{0xf3, rexPrefixW, 0xa5}},
+		{inst: REPSTOSQ, exp: []byte{0xf3, rexPrefixW, 0xab}},
 		{inst: STD, exp: []byte{0xfd}},
 		{inst: CLD, exp: []byte{0xfc}},
 	}