```
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: AMD Ryzen 9 3950X 16-Core Processor
                                    │   old.txt   │              new.txt              │
                                    │   sec/op    │   sec/op     vs base              │
Compilation/without_extern_cache-32   4.441m ± 3%   4.034m ± 1%  -9.16% (p=0.001 n=7)
Compilation_sqlite3/compiler-32       517.0m ± 3%   478.6m ± 2%  -7.42% (p=0.001 n=7)
geomean                               47.91m        43.94m       -8.29%

                                    │   old.txt    │              new.txt               │
                                    │     B/op     │     B/op      vs base              │
Compilation/without_extern_cache-32   665.0Ki ± 0%   665.0Ki ± 0%       ~ (p=0.557 n=7)
Compilation_sqlite3/compiler-32       49.46Mi ± 0%   49.46Mi ± 0%       ~ (p=0.535 n=7)
geomean                               5.668Mi        5.668Mi       -0.00%

                                    │   old.txt   │               new.txt               │
                                    │  allocs/op  │  allocs/op   vs base                │
Compilation/without_extern_cache-32    925.0 ± 0%    925.0 ± 0%       ~ (p=1.000 n=7) ¹
Compilation_sqlite3/compiler-32       36.99k ± 0%   36.99k ± 0%       ~ (p=0.805 n=7)
geomean                               5.850k        5.850k       +0.00%
```